Extract methods previously in Duke to respective classes (Ui, Storage or Parser)
