+++ data/tasks.txt
@@ -1,2 +1,1 @@
-D|false|test|02/02/2023
-E|false|test|02/02/2023|03/02/2023
+T|false|testtodo
