+++ text-ui-test/EXPECTED.TXT
@@ -20,7 +20,6 @@
 	____________________________________________________________
 	Here are the tasks in your list:
     1. [T][ ]testtodo
-    testerror
 	____________________________________________________________
 	____________________________________________________________
 	Bye! Hope to see you again soon!
